[fix, rename, chore] useMutation 커스텀 훅 onError 로직 수정, clientFetchers 네이밍 변경, react-datepicker 제거
