+++ package.json
@@ -32,7 +32,6 @@
     "motion": "^12.11.2",
     "next": "15.3.2",
     "react": "^19.0.0",
-    "react-datepicker": "^8.3.0",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.56.4",
     "tailwind-merge": "^3.3.0",
