+++ src/app/api/gatherings/participants/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { EXTERNAL_PATHS } from '@/lib/api/apiPaths';
 import { AxiosError } from 'axios';
-import { apiServer } from '@/lib/api/clientFetcher';
+import { apiServer } from '@/lib/api/clientFetchers';
 
 /**
  * 모임 참여자 목록 조회
