+++ src/components/gatherings/shared/utils/fetchPaginatedGatherings.ts
@@ -1,5 +1,5 @@
 import { INTERNAL_PATHS } from '@/lib/api/apiPaths';
-import { apiClient } from '@/lib/api/clientFetcher';
+import { apiClient } from '@/lib/api/clientFetchers';
 import { Gathering } from "@/types/gatherings";
 
 /**
