+++ src/hooks/api/gatherings/detail/useLeaveGathering.ts
@@ -2,7 +2,7 @@
 
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { GatheringApiParams } from '@/types/gatheringApi';
-import { apiClient } from '@/lib/api/clientFetcher';
+import { apiClient } from '@/lib/api/clientFetchers';
 import { INTERNAL_PATHS } from '@/lib/api/apiPaths';
 import { handleApiError } from '@/lib/api/handleApiResponse';
 
@@ -28,7 +28,15 @@ export const useLeaveGathering = ({ token, onCallback }: GatheringApiParams) => 
         },
         onError: (error) => {
             const response = handleApiError(error);
-            response.text().then(message => onCallback?.(message));
+            response.text().then(text => {
+                try {
+                    const json = JSON.parse(text);
+                    const message = json?.error?.message || json?.message || text;
+                    onCallback?.(message);
+                } catch {
+                    onCallback?.(text);
+                }
+            });
         }
     });
