+++ src/hooks/api/mypage/useCreateReview.ts
@@ -2,7 +2,7 @@
 
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { INTERNAL_PATHS } from '@/lib/api/apiPaths';
-import { apiClient } from '@/lib/api/clientFetcher';
+import { apiClient } from '@/lib/api/clientFetchers';
 import { GatheringApiParams } from '@/types/gatheringApi';
 import { handleApiError } from '@/lib/api/handleApiResponse';
 
@@ -33,7 +33,15 @@ export const useCreateReview = ({ token, onCallback }: GatheringApiParams) => {
         },
         onError: (error) => {
             const response = handleApiError(error);
-            response.text().then(message => onCallback?.(message));
+            response.text().then(text => {
+                try {
+                    const json = JSON.parse(text);
+                    const message = json?.error?.message || json?.message || text;
+                    onCallback?.(message);
+                } catch {
+                    onCallback?.(text);
+                }
+            });
         }
     });
